update is_enabled() call and build nvim script

+++ utils/build_nvim.sh
@@ -2,6 +2,15 @@
 
 set -e
 
+if [ "$(basename $(pwd))" != "neovim" ]; then
+  echo "[!] probably wrong directory, exiting just in case"
+  exit 1
+fi
+
+echo "[*] pulling latest changes"
+git pull
+
+echo "[*] building newest version and installing it"
 make distclean \
   && make clean \
   && rm -rf build out \
